move back into util package, and make package-private. Some minor renames to use a common prefix across types
